Return integer dispatch quantities

+++ backend/IQBF.Application/DTOs/Dispatches/DispatchDto.cs
@@ -1,10 +1,6 @@
 namespace IQBF.Application.DTOs.Dispatches;
 
-public sealed record DispatchItemDto(
-    Guid BLId,
-    string BLCode,
-    decimal Quantity
-);
+public sealed record DispatchItemDto(Guid BLId, string BLCode, int Quantity);
 
 public sealed record DispatchDto(
     Guid Id,
